+++ docker-compose/acceptance_test.sh
@@ -4,4 +4,5 @@ set -Ceu -o pipefail
 #set -vx # for debug
 
 MYSQL_VERSION=${1:-mysql80}
-docker-compose -f docker-compose.yml -f "docker-compose/$MYSQL_VERSION.yml" run -e "MYSQL_HOST=$MYSQL_VERSION" app acceptance/run_test.sh
+DOCKER_COMPOSE_FILE=${2:-docker-compose.yml}
+docker-compose -f "$DOCKER_COMPOSE_FILE" -f "docker-compose/$MYSQL_VERSION.yml" run -e "MYSQL_HOST=$MYSQL_VERSION" app acceptance/run_test.sh