+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `masking --version` option [#31](https://github.com/kibitan/masking/pull/31)
+
 ## [v0.0.3] - 2019-07-07
 
 ### Changed