+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `masking --version` option [#31](https://github.com/kibitan/masking/pull/31)
 - add friendly parse error message [#26](https://github.com/kibitan/masking/pull/26)
 - Setup Docker and Acceptance test [#32](https://github.com/kibitan/masking/pull/32)
+- setup acceptance test for ci [#33](https://github.com/kibitan/masking/pull/33)
 
 ## [v0.0.3] - 2019-07-07