+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `masking --version` option [#31](https://github.com/kibitan/masking/pull/31)
-- add friendly parse error message [#26](https://github.com/kibitan/masking/pull/26)
+- add friendly parse error message [#26](https://github.com/kibitan/masking/pull/26) [#38](https://github.com/kibitan/masking/pull/38)
 - Setup Docker and Acceptance test [#32](https://github.com/kibitan/masking/pull/32) [#36](https://github.com/kibitan/masking/pull/36)
 - setup acceptance test for ci [#34](https://github.com/kibitan/masking/pull/34)
 - add Ruby2.7(preview) [#35](https://github.com/kibitan/masking/pull/35)