+++ CHANGELOG.md
@@ -46,7 +46,8 @@ $ bin/benchmark.rb
 
 Initial release version. 🎉
 
-[Unreleased]: https://github.com/kibitan/masking/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/kibitan/masking/compare/v1.0.0...HEAD
+[v1.0.0]: https://github.com/kibitan/masking/compare/v0.0.3...v1.0.0
 [v0.0.3]: https://github.com/kibitan/masking/compare/v0.0.2...v0.0.3
 [v0.0.2]: https://github.com/kibitan/masking/compare/v0.0.1...v0.0.2
 [v0.0.1]: https://github.com/kibitan/masking/tree/v0.0.1